actualizando migraciones

+++ Expensive.Persistance/Migrations/ExpensiveApplicationDataContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace Expensive.Persistance.Migrations
 
             modelBuilder.Entity("Expensive.Domain.Entities.Categories", b =>
                 {
-                    b.Property<int>("CategorieId")
+                    b.Property<long>("CategorieId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("CategorieId"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("CategorieId"));
 
                     b.Property<DateTime>("CreatedAt")
                         .ValueGeneratedOnAdd()
@@ -99,7 +99,7 @@ namespace Expensive.Persistance.Migrations
                     b.HasData(
                         new
                         {
-                            CategorieId = 1,
+                            CategorieId = 1L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de gastos del Hogar",
@@ -112,7 +112,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            CategorieId = 2,
+                            CategorieId = 2L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de gastos del Automovil",
@@ -125,7 +125,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            CategorieId = 3,
+                            CategorieId = 3L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de gastos del Mascotas",
@@ -138,7 +138,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            CategorieId = 4,
+                            CategorieId = 4L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de gastos del Salud",
@@ -151,7 +151,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            CategorieId = 5,
+                            CategorieId = 5L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de gastos del Mercado",
@@ -164,7 +164,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            CategorieId = 6,
+                            CategorieId = 6L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de gastos del Entretenimiento",
@@ -177,7 +177,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            CategorieId = 7,
+                            CategorieId = 7L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de ingresos del ahorro",
@@ -190,7 +190,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            CategorieId = 8,
+                            CategorieId = 8L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Categoria de ingresos del sueldo",
@@ -205,11 +205,11 @@ namespace Expensive.Persistance.Migrations
 
             modelBuilder.Entity("Expensive.Domain.Entities.PaymentMethods", b =>
                 {
-                    b.Property<int>("PaymentMethodId")
+                    b.Property<long>("PaymentMethodId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("PaymentMethodId"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("PaymentMethodId"));
 
                     b.Property<DateTime>("CreatedAt")
                         .ValueGeneratedOnAdd()
@@ -275,7 +275,7 @@ namespace Expensive.Persistance.Migrations
                     b.HasData(
                         new
                         {
-                            PaymentMethodId = 1,
+                            PaymentMethodId = 1L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Metodo de pago en efectivo",
@@ -287,7 +287,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            PaymentMethodId = 2,
+                            PaymentMethodId = 2L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Metodo de pago en Tarjeta de Credito",
@@ -299,7 +299,7 @@ namespace Expensive.Persistance.Migrations
                         },
                         new
                         {
-                            PaymentMethodId = 3,
+                            PaymentMethodId = 3L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             Description = "Metodo de pago en Tarjeta de Debito",
@@ -319,8 +319,8 @@ namespace Expensive.Persistance.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("UserCategorieId"));
 
-                    b.Property<int>("CategorieId")
-                        .HasColumnType("int");
+                    b.Property<long>("CategorieId")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("CreatedAt")
                         .ValueGeneratedOnAdd()
@@ -356,8 +356,8 @@ namespace Expensive.Persistance.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<int>("UserId")
-                        .HasColumnType("int");
+                    b.Property<long>("UserId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("UserCategorieId");
 
@@ -378,100 +378,100 @@ namespace Expensive.Persistance.Migrations
                         new
                         {
                             UserCategorieId = 1L,
-                            CategorieId = 1,
+                            CategorieId = 1L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
                             UserCategorieId = 2L,
-                            CategorieId = 2,
+                            CategorieId = 2L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
                             UserCategorieId = 3L,
-                            CategorieId = 3,
+                            CategorieId = 3L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
                             UserCategorieId = 4L,
-                            CategorieId = 4,
+                            CategorieId = 4L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
                             UserCategorieId = 5L,
-                            CategorieId = 5,
+                            CategorieId = 5L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
                             UserCategorieId = 6L,
-                            CategorieId = 6,
+                            CategorieId = 6L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
                             UserCategorieId = 7L,
-                            CategorieId = 7,
+                            CategorieId = 7L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
                             UserCategorieId = 8L,
-                            CategorieId = 8,
+                            CategorieId = 8L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         });
                 });
 
             modelBuilder.Entity("Expensive.Domain.Entities.UserPaymentMethods", b =>
                 {
-                    b.Property<int>("UserPaymentMethodId")
+                    b.Property<long>("UserPaymentMethodId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("UserPaymentMethodId"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("UserPaymentMethodId"));
 
                     b.Property<DateTime>("CreatedAt")
                         .ValueGeneratedOnAdd()
@@ -493,8 +493,8 @@ namespace Expensive.Persistance.Migrations
                         .HasColumnType("bit")
                         .HasDefaultValue(false);
 
-                    b.Property<int>("PaymentMethodId")
-                        .HasColumnType("int");
+                    b.Property<long>("PaymentMethodId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("RowGuid")
                         .ValueGeneratedOnAdd()
@@ -510,8 +510,8 @@ namespace Expensive.Persistance.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<int>("UserId")
-                        .HasColumnType("int");
+                    b.Property<long>("UserId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("UserPaymentMethodId");
 
@@ -531,46 +531,46 @@ namespace Expensive.Persistance.Migrations
                     b.HasData(
                         new
                         {
-                            UserPaymentMethodId = 1,
+                            UserPaymentMethodId = 1L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
-                            PaymentMethodId = 1,
+                            PaymentMethodId = 1L,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
-                            UserPaymentMethodId = 2,
+                            UserPaymentMethodId = 2L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
-                            PaymentMethodId = 2,
+                            PaymentMethodId = 2L,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         },
                         new
                         {
-                            UserPaymentMethodId = 3,
+                            UserPaymentMethodId = 3L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "Created By Default",
                             IsActive = false,
                             IsDeleted = false,
-                            PaymentMethodId = 3,
+                            PaymentMethodId = 3L,
                             UpdatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
-                            UserId = 1
+                            UserId = 1L
                         });
                 });
 
             modelBuilder.Entity("Expensive.Domain.Entities.Users", b =>
                 {
-                    b.Property<int>("UserId")
+                    b.Property<long>("UserId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("UserId"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("UserId"));
 
                     b.Property<DateTime>("CreatedAt")
                         .ValueGeneratedOnAdd()
@@ -651,7 +651,7 @@ namespace Expensive.Persistance.Migrations
                     b.HasData(
                         new
                         {
-                            UserId = 1,
+                            UserId = 1L,
                             CreatedAt = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             CreatedBy = "System",
                             Email = "miguelmodd@gmail.com",
